requirements.txt
pandas
matplotlib
Pillow
ftfy
regex
tqdm
opencv-python
PyYAML
ultralytics
gdown
torch-tb-profiler
scikit-learn
requests
litellm
pyarrow
torchvision
wandb
itakello_logging
git+https://github.com/openai/CLIP.git
flake8
black
mypy
ollama
pytest